services/ranking_service/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
httpx
redis
